# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Climate Action Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get gamified leaderboard
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: category
          in: query
          required: false
          description: Filter to a mission category leaderboard. Omit for the all-time overall leaderboard.
          schema:
            type: string
            enum: [daily, weekly, monthly, awareness]
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /leaderboard/me:
    get:
      operationId: getMyRank
      tags: [leaderboard]
      summary: Get the current user's rank and points
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [daily, weekly, monthly, awareness]
      responses:
        "200":
          description: User rank info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardEntry"
...
components:
  schemas:
...
    LeaderboardEntry:
      type: object
      required: [rank, userId, name, points, completions, level]
      properties:
        rank:
          type: integer
        userId:
          type: integer
        name:
          type: string
        avatarUrl:
          type: ["string", "null"]
        points:
          type: integer
        completions:
          type: integer
        level:
          type: string
        badge:
          type: ["string", "null"]
        title:
          description: Marvel-themed title awarded to the top 5 ranks of a category leaderboard
          type: ["string", "null"]
        isStoneCollector:
          description: True if this user holds rank #1 in all four category leaderboards at once
          type: boolean
...